+++ src/Plugin/Wechat/Pay/Combine/QueryRefundPlugin.php
@@ -7,6 +7,6 @@ namespace Yansongda\Pay\Plugin\Wechat\Pay\Combine;
 /**
  * @see https://pay.weixin.qq.com/wiki/doc/apiv3/apis/chapter5_1_15.shtml
  */
-class FindRefundPlugin extends \Yansongda\Pay\Plugin\Wechat\Pay\Common\FindRefundPlugin
+class QueryRefundPlugin extends \Yansongda\Pay\Plugin\Wechat\Pay\Common\QueryRefundPlugin
 {
 }