+++ src/Plugin/Wechat/Pay/Native/QueryRefundPlugin.php
@@ -7,6 +7,6 @@ namespace Yansongda\Pay\Plugin\Wechat\Pay\Native;
 /**
  * @see https://pay.weixin.qq.com/wiki/doc/apiv3/apis/chapter3_4_10.shtml
  */
-class FindRefundPlugin extends \Yansongda\Pay\Plugin\Wechat\Pay\Common\FindRefundPlugin
+class QueryRefundPlugin extends \Yansongda\Pay\Plugin\Wechat\Pay\Common\QueryRefundPlugin
 {
 }