+++ src/Plugin/Wechat/Shortcut/QueryShortcut.php
@@ -7,8 +7,8 @@ namespace Yansongda\Pay\Plugin\Wechat\Shortcut;
 use Yansongda\Pay\Contract\ShortcutInterface;
 use Yansongda\Pay\Exception\Exception;
 use Yansongda\Pay\Exception\InvalidParamsException;
-use Yansongda\Pay\Plugin\Wechat\Pay\Common\FindRefundPlugin;
 use Yansongda\Pay\Plugin\Wechat\Pay\Common\QueryPlugin;
+use Yansongda\Pay\Plugin\Wechat\Pay\Common\QueryRefundPlugin;
 use Yansongda\Supports\Str;
 
 class QueryShortcut implements ShortcutInterface
@@ -41,7 +41,7 @@ class QueryShortcut implements ShortcutInterface
     protected function refundPlugins(): array
     {
         return [
-            FindRefundPlugin::class,
+            QueryRefundPlugin::class,
         ];
     }